requirements.txt
motor
PyJWT
bcrypt
uvicorn
pymongo
python-decouple
fastapi
pydantic[email]
passlib
loguru
aioredis
hiredis
fastapi-cache2